// aiot-frontend/src/api/generated/device-service/models/ota-v1-create-otapackage-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * CreateOTAPackage
 */
export interface OtaV1CreateOTAPackageRequest {
    'packageName'?: string;
    'version'?: string;
    'productId'?: string;
    'packageType'?: string;
    'uploadType'?: string;
    'fileUrl'?: string;
    'fileSize'?: string;
    'checksum'?: string;
    'description'?: string;
    'releaseNotes'?: string;
    'metadata'?: string;
}